refactor(db): use dynamic import for subcommand
