+++ packages/opencode/src/index.ts
@@ -29,7 +29,7 @@ import { EOL } from "os"
 import { WebCommand } from "./cli/cmd/web"
 import { PrCommand } from "./cli/cmd/pr"
 import { SessionCommand } from "./cli/cmd/session"
-import { DbCommand } from "./cli/cmd/db"
+import { DbCommand } from "./cli/cmd/db/command"
 import path from "path"
 import { Global } from "@opencode-ai/core/global"
 import { JsonMigration } from "@/storage/json-migration"
